Feat: Add fetched data from tables to XLS workbook
Save collected data in a given workbook and for the given sheetname,
according to the type of the collected data, it means, if its an array
of arrays or an array of objects. To process this data use Sheet JS
library. Also, if worksheet already exists, append given data to the
below end address cell of the contents on it.

modified:   src/js/xls-output.js